fix: timeouting ipv6 tests on some machines fix
When configuring network namespace with IPv6 addresses, on some
machines when sending requests immediately after configuring netns
they are time-outing, when on others there is 1-2s delay before
receiving the response.
I was able to narrow down the cause to be related to
IPv6 neighbor discovery, but was not able to find a real reason
or a fix, so we have to introduce this delay.
The other observation I got is when sending requests
immediately, they are going through ens5/eth0 interface instead of
created by us veth main one

Signed-off-by: Bart Smykla <bartek@smykla.com>